# perf(es/minifier): Speed up `merge_sequences_in_exprs` by caching computation