--- a/.changeset/forty-maps-smoke.md
+++ b/.changeset/forty-maps-smoke.md
@@ -1,0 +1,6 @@
+---
+swc_core: patch
+swc_ecma_minifier: patch
+---
+
+perf(es/minifier): Speed up `merge_sequences_in_exprs` by caching computation